Update SIGMET and AIRMET stationary movement
Align the documentation of directionOfMotion and speedOfMotion with TAC-to-XML-Guidance.txt.
Specific values should report the stationary movement. It allows for distinguishing between the stationary movement and the case when no movement is reported. The case when no movement is reported can occur when SIGMET is created from advisory, and the observation does not affect the FIR; only the +6-hour forecast affects the FIR. The movement is forecasted only for the observation in the advisory.

Related to #250